+++ lib/Perl/Tidy.pm
@@ -132,7 +132,7 @@ BEGIN {
     # then the Release version must be bumped, and it is probably past time for
     # a release anyway.
 
-    $VERSION = '20230912.03';
+    $VERSION = '20230912.04';
 } ## end BEGIN
 
 sub DESTROY {