+++ lib/Perl/Tidy/Tokenizer.pm
@@ -33,7 +33,7 @@ use strict;
 use warnings;
 use English qw( -no_match_vars );
 
-our $VERSION = '20230912.03';
+our $VERSION = '20230912.04';
 
 use Carp;